Wow it actually works
