+++ src/renderer/actions/types.ts
@@ -1,0 +1,8 @@
+import { ActionTypeGenerator } from "@vinceau/event-actions";
+import { FC } from "react";
+
+export interface ActionComponent {
+    label: string;
+    action: ActionTypeGenerator;
+    Component: FC;
+}
